Refactor the bloom into a modifier.
It currently depends on the home screen context but that will be removed in the future.
